requirements.txt
boto3==1.42.91
fastapi==0.115.0
uvicorn==0.30.6
python-dotenv==1.0.1
pydantic==2.9.2
langchain-aws>=0.1.16
langchain>=0.1.0
langchain-core>=0.1.0
requests>=2.31.0
pytz>=2024.1
cognee>=1.0.0
fastembed>=0.8.0
s3fs>=2026.3.0
selenium>=4.20.0
beautifulsoup4>=4.12.0
pymupdf>=1.24.0
langchain-text-splitters>=0.2.0
